+++ packages/insights-charts/source/js/helpers/getLocale.js
@@ -1,0 +1,11 @@
+const getLocale = () => {
+  const navigator = window.navigator;
+
+  if (navigator.languages) {
+    return navigator.languages[0];
+  }
+
+  return navigator.language;
+};
+
+export default getLocale;